Added test that checks the counter correctly increases.
Each DI can be triggered which will increament the DI's associated
counter. Typically we want these counter values to remain persistent
even on restart (as they are a safety feature). The resetDICounter
function allows for these to be reset, but add considerable time to
the test length, as such it shouldn't be in setUP.