Add test for high actual temp behaviour
This isn't present on the flowchart. Actually added in by the script v3
update. There was an issue where the system was unable to cool if the
actual temp was high above the cutoff. Now takes both into account. Now
takes both the sp and the actual temp into account.